Add files via upload
The objective of this code is to perform a series of data management and analysis tasks related to COVID-19 statistics within a SQL database. It starts by renaming the database and then proceeds to extract and manipulate data from tables. The code calculates various statistics, including death percentages, infection rates, and global COVID-19 statistics, with the goal of storing relevant data for visualisation. This code facilitates the extraction of insights and visualisations from the COVID-19 data, aiding in the understanding of the pandemic's impact and trends.